update: [Refactor] SinkState in separate file
